src/requirements.txt
matplotlib==3.8.2
numpy==1.26.3
pillow==10.2.0